% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_qualities.R
\name{decode_qualities}
\alias{decode_qualities}
\title{Converts quality values to Phred score}
\usage{
decode_qualities(qualities, offset = 33)
}
\arguments{
\item{qualities}{A character vector of qualities}

\item{offset}{Must be a value of 33 or 64}
}
\value{
A vector of Phred scores
}
\description{
Converts a scalar string of quality values into a vector of Phred scores.
}
\examples{
decode_qualities("???#;ABAAAH")
}
