% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package docs.R
\docType{package}
\name{fastqr}
\alias{fastqr}
\title{fastqr: A package for read in fastq files and converting their qualities to
Phred scores.}
\description{
The fastqr package provides three main functions which you might
want to use.
}
\section{sangr functions}{

gc_content

read_fastq

decode_qualities
}

