% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq.R
\name{gc_content}
\alias{gc_content}
\title{\% of GC values in a DNA sequence}
\usage{
gc_content(seq)
}
\arguments{
\item{seq}{A vector of DNA sequence strings}
}
\value{
Percentage of bases that are G or C as a vector.
}
\description{
Provides the percentage of bases with values of G or C in a DNA sequence.
}
\examples{
gc_content("GAGAGCGGCTT")
}
