% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq.R
\name{read_fastq}
\alias{read_fastq}
\title{Reads fastq file and puts into tibble}
\usage{
read_fastq(file)
}
\arguments{
\item{file}{The fastq file to read from. Requires extension of .fq.}
}
\value{
A tibble with 4 columns. \cr
1) ID: The sequence from the first line minus the "@" \cr
2) Bases: The bases from the second line \cr
3) Qualities: The quality string from the fourth line \cr
4) GC: The GC content of the bases \cr
}
\description{
Reads a fastq file and puts into a tibble with one row per fastq entry.
}
\examples{
read_fastq(system.file("good.fq", package = "fastqR"))
}
