% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_qualities.R
\name{decode_qualities}
\alias{decode_qualities}
\title{Decode qualities}
\usage{
decode_qualities(qualities, offset = 33)
}
\arguments{
\item{qualities}{set of quality scores from a fastq file}

\item{offset}{phred type (numerical) must be 33 or 64}
}
\value{
numeric vector of Phred scores
}
\description{
Decode quality scores from fastq file.
}
\examples{
decode_qualities("???#;ABAAAH")
}
