% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_qualities.R
\name{decode_qualities}
\alias{decode_qualities}
\title{Title}
\usage{
decode_qualities(qualities, offset = 33)
}
\arguments{
\item{qualities}{String of sequence Qualities from a fastq format file}

\item{offset}{Scalar number which is either 33 or 64, the only two offsets used in fastq files. Default value is 33}
}
\value{
Vector of phred scores for each character of the sequence qualities, values cannot be < 0
}
\description{
Title
}
\examples{
decode_qualities("???#;ABAAAH")
}
