% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_qualities.R
\name{decode_qualities}
\alias{decode_qualities}
\title{Title Decodes a FastQ quality string into  scores}
\usage{
decode_qualities(qualities, offset = 33)
}
\arguments{
\item{qualities}{A sfastq quality string}

\item{offset}{The Phred encoding offset value}
}
\value{
A numeric vector of Phred scores
}
\description{
Title Decodes a FastQ quality string into  scores
}
