% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_qualities.R
\name{decode_qualities}
\alias{decode_qualities}
\title{Decode FastQ quality strings into Phred scores}
\usage{
decode_qualities(qualities, offset = 33)
}
\arguments{
\item{qualities}{a string scalar containing FastQ encoded quality scores}

\item{offset}{the ASCII offset (33 or 64) of the encoded phred scores}
}
\value{
a vector of Phred scores
}
\description{
Decode FastQ quality strings into Phred scores
}
\examples{
decode_qualities("???#;ABAAAH")

decode_qualities("WZZZVX[]", offset=64)
}
