% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NULL.R
\docType{package}
\name{fastqR}
\alias{fastqR}
\title{fastqR: A package for importing and interpreting fastq files}
\description{
The fastqR package contains functions to import fastq files, to decode quality
scores and to calculate GC content of reads.
}
\section{fastqR functions}{

read_fastq

gc_content

decode_qualities
}

