% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{fastqR}
\alias{fastqR}
\title{fastqR: A package for reading and manipulating fastq format sequence files}
\description{
The fastqR package provides three main functions which you might
want to use.
}
\section{fastqR functions}{

read_fastq

gc_content

decode_qualities
}

