% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqR_documentation.R
\docType{package}
\name{fastqR}
\alias{fastqR}
\title{fastqR: A package to read and manipulate FastQ files}
\description{
This package works with FastQ format files.  It can read these
into a tidyverse compatible data structure and provides
functions for manipulating the data within them.
}
