% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq.R
\name{gc_content}
\alias{gc_content}
\title{Calculate GC content of nucleotide sequence}
\usage{
gc_content(seq)
}
\arguments{
\item{seq}{A vector containing strings of nucleotides to use}
}
\value{
Vector of percentage values of sequence GC contents
}
\description{
Calculate GC content of nucleotide sequence
}
\examples{
gc_content("GAGAGCGGCTT")
}
