% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq.R
\name{gc_content}
\alias{gc_content}
\title{Title Calculates the QC content of a set of DNA sequences}
\usage{
gc_content(seq)
}
\arguments{
\item{seq}{A character vestor of DNA sequences}
}
\value{
A numeric vector of GC percentages
}
\description{
Title Calculates the QC content of a set of DNA sequences
}
