% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq.R
\name{gc_content}
\alias{gc_content}
\title{Calculate DNA GC content}
\usage{
gc_content(seq)
}
\arguments{
\item{seq}{a character vector of DNA sequences}
}
\value{
a vector of percentage GC values can be upper or lower case. Any non GATC characters will be treated as non-GC
}
\description{
Calculate DNA GC content
}
\examples{
gc_content(c("GGATCG","gtattagc"))
}
