% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq.R
\name{read_fastq}
\alias{read_fastq}
\title{Read fastq file}
\usage{
read_fastq(filename)
}
\arguments{
\item{filename}{fastq file to read in}
}
\value{
tibble with columns containing the id, sequence, id2 and quality score.
Each read is on a new row.
}
\description{
Read fastq file
}
\examples{
fq_file <- read_fastq(system.file("good.fq", package = "fastqR"))
}
