% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq.R
\name{read_fastq}
\alias{read_fastq}
\title{Read fastq .fq file and calculate GC content}
\usage{
read_fastq(file)
}
\arguments{
\item{file}{A .fq format file}
}
\value{
A tibble with with ID, Bases, Qualities and GC content of the sequences in the fastq file
}
\description{
Read fastq .fq file and calculate GC content
}
\examples{
read_fastq(system.file("good.fq", package = "fastqR"))
}
