% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fastq.R
\name{read_fastq}
\alias{read_fastq}
\title{Read in a FastQ file}
\usage{
read_fastq(file)
}
\arguments{
\item{file}{an uncompressed file (.fq extension) in fastq format}
}
\value{
A tibble with ID, Bases, Qualities and GC columns
}
\description{
Read in a FastQ file
}
\examples{
read_fastq(system.file("good.fq", package = "fastqR"))
}
